+++ pkg/diag/errors.go
@@ -72,6 +72,11 @@ func GetCannotDeleteParentResourceWithoutAlsoDeletingChildError(urn resource.URN
 	return newError(urn, 2012, "Cannot delete parent resource '%v' without also deleting child '%v'.")
 }
 
-func GetResourceIsBeingCreatedButWasNotSpecifiedInTargetList(urn resource.URN) *Diag {
-	return newError(urn, 2013, "Resource '%v' is being created but was not specified in -target list.")
+func GetResourceWillBeCreatedButWasNotSpecifiedInTargetList(urn resource.URN) *Diag {
+	return newError(urn, 2013, `Resource '%v' depends on '%v' which was was not specified in --target list.`)
+}
+
+func GetResourceWillBeDestroyedButWasNotSpecifiedInTargetList(urn resource.URN) *Diag {
+	return newError(urn, 2014, `Resource '%v' will be destroyed but was not specified in --target list.
+Either include resource in --target list or pass --target-dependents to proceed.`)
 }